Rename noopcf to noopservicescontroller
[#151720679]
