+++ cmd/on-demand-service-broker/main.go
@@ -24,7 +24,7 @@ import (
 	"github.com/pivotal-cf/on-demand-service-broker/cf"
 	"github.com/pivotal-cf/on-demand-service-broker/config"
 	"github.com/pivotal-cf/on-demand-service-broker/loggerfactory"
-	"github.com/pivotal-cf/on-demand-service-broker/noopcf"
+	"github.com/pivotal-cf/on-demand-service-broker/noopservicescontroller"
 	"github.com/pivotal-cf/on-demand-service-broker/serviceadapter"
 	"github.com/pivotal-cf/on-demand-service-broker/task"
 )
@@ -92,7 +92,7 @@ func startBroker(conf config.Config, logger *log.Logger, loggerFactory *loggerfa
 			logger.Fatalf("error creating Cloud Foundry client: %s", err)
 		}
 	} else {
-		cfClient = noopcf.New()
+		cfClient = noopservicescontroller.New()
 	}
 
 	serviceAdapter := &serviceadapter.Client{
