+++ noopservicescontroller/noopservicescontroller_suite_test.go
@@ -1,4 +1,4 @@
-package noopcf_test
+package noopservicescontroller_test
 
 import (
 	. "github.com/onsi/ginkgo"
